+++ pyccl/baryons/__init__.py
@@ -1,0 +1,8 @@
+from .baryons_base import Baryons
+from .schneider15 import BaryonsSchneider15
+
+
+__all__ = (
+    "Baryons",
+    "BaryonsSchneider15"
+)
